requirements.txt
tqdm==4.40.0
numpy==1.16.0
setuptools==42.0.2
torch==1.5.0
typing==3.7.4.1
pandas==0.24.0
falcon==2.0.0
kafka-python==1.4.6
gunicorn==20.0.4
redis==3.5.0
requests==2.23.0
urllib3==1.25.9
pymongo==3.10.1
mongoengine==0.20.0
faiss-cpu==1.6.3
statsd==3.3.0